Entity and Skill models moved to shared app
