+++ resumes/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
 
-from resumes.models import Resume
+from .models import Resume
 
 admin.site.register(Resume)
